validate input util
